+++ substudy/src/export/tracks.rs
@@ -9,6 +9,7 @@ use log::debug;
 use crate::{
     export::Exporter,
     time::{seconds_to_hhmmss, Period},
+    ui::Ui,
     video::Id3Metadata,
     Result,
 };
@@ -46,7 +47,7 @@ struct Conv {
 }
 
 /// Export the video as a set of tracks.
-pub fn export_tracks(exporter: &mut Exporter) -> Result<()> {
+pub fn export_tracks(ui: &Ui, exporter: &mut Exporter) -> Result<()> {
     let mut convs: Vec<Conv> = vec![];
 
     // Figure out how to combine subtitles into conversations.
@@ -132,7 +133,7 @@ pub fn export_tracks(exporter: &mut Exporter) -> Result<()> {
     exporter.export_data_file("playlist.m3u8", &playlist)?;
 
     // Extract our media files.
-    exporter.finish_exports()?;
+    exporter.finish_exports(ui)?;
 
     Ok(())
 }