+++ substudy/src/import/mod.rs
@@ -1,5 +1,5 @@
 //! Functions for importing data from various sources.
 
-pub use self::whisper::import_whisper_json;
+pub use self::whisper::{import_whisper_json, WhisperJson};
 
 mod whisper;