+++ substudy/src/services/oai/mod.rs
@@ -15,8 +15,15 @@ use tokio::time::sleep;
 
 use crate::Result;
 
-pub use self::translate::translate_subtitle_file;
+pub use self::{
+    transcribe::{
+        transcribe_subtitles_to_srt_file, transcribe_subtitles_to_whisper_json,
+        TranscriptionFormat,
+    },
+    translate::translate_subtitle_file,
+};
 
+mod transcribe;
 mod translate;
 
 /// Retry an OpenAI request a few times.